Moved process to service layer
